Fix date parsing to get date string more accurately.
Previously, `todays_date` in the `remote-backup-mysql.sh` script tried
to cut out the date from a date and time string using the `cut` utility.
This lead to errors when the format was not exactly as expected. To fix
this, we can just let `date` display the fields we need and then output
in the required format.